refactor(docs): card dx
