+++ apps/docs/content/components/card/usage.raw.jsx
@@ -1,0 +1,11 @@
+import {Card, CardBody} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Card>
+      <CardBody>
+        <p>Make beautiful websites regardless of your design experience.</p>
+      </CardBody>
+    </Card>
+  );
+}
