+++ apps/docs/content/components/card/usage.ts
@@ -1,14 +1,4 @@
-const App = `import {Card, CardBody} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Card>
-      <CardBody>
-        <p>Make beautiful websites regardless of your design experience.</p>
-      </CardBody>
-    </Card>
-  );
-}`;
+import App from "./usage.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
